restructured, added self import
