+++ src/python_algebra/math_tree.py
