+++ tests/test_math_tree.py
@@ -4,14 +4,14 @@ Unittests for math_tree using pytest
 
 from typing import Callable
 
-from hypothesis import given, settings
+from hypothesis import given
 from hypothesis.strategies import booleans, builds, deferred, dictionaries, floats, integers, one_of, \
     sampled_from, SearchStrategy
-from math_tree import Absolute, And, ArbitraryOperator, ArcCosine, ArcSine, ArcTangent, Cosine, Derivative, Division, \
-    E, EvaluationError, Exponent, GreaterEqual, GreaterThan, Integer, Invert, IsEqual, LessEqual, LessThan, Logarithm, \
-    Nand, Negate, Node, Nodeify, Nor, Not, NotEqual, Or, Pi, Piecewise, Product, Sine, Subtraction, Sum, Tangent, \
-    UnaryOperator, Variable, Environment, Xnor, Xor
-from pytest import fixture, raises
+from python_algebra.math_tree import Absolute, And, ArbitraryOperator, ArcCosine, ArcSine, ArcTangent, Cosine, \
+    Derivative, Division, E, EvaluationError, Exponent, GreaterEqual, GreaterThan, Integer, Invert, IsEqual, LessEqual,\
+    LessThan, Logarithm, Nand, Negate, Node, Nodeify, Nor, Not, NotEqual, Or, Pi, Piecewise, Product, Sine,\
+    Subtraction, Sum, Tangent, UnaryOperator, Variable, Environment, Xnor, Xor
+from pytest import fixture
 
 
 @fixture(scope="module")
